Updates to make commands pipeable #1.
OK I have made changes so that you can pipe these commands.

You should now be able to do things like.

```powershell
> Vsix-IncrementVsixVersion | Vsix-UpdateBuildVersion

> '.\source.extension.vsixmanifest' | Vsix-IncrementVsixVersion |
> Vsix-UpdateBuildVersion

>@('.\source.extension.vsixmanifest','.\source2.extension.vsixmanifest') |
>Vsix-IncrementVsixVersion | Vsix-UpdateBuildVersion -version 1.1
```

Roughly what I did was add a process block and the ValueFromPipeline on
the parameter for the vsix file path. To support pipelining the functions
need to return values to the pipeline so in some cases you'll see that
I've added a statement that outputs the filename to the pipleine. When
pipelining you should be careful with returing any extra results so I
added a couple of | Out-Null to the end of some calls.

I also changed some lines from Write-Host 'sample' to 'sample' |
Write-Host that is my preferred way to call Write-Host and I recommend
that for you as well.

I couldn't test the Vsix-PublishToGallery locally so try it in AppVeyor
and let me know if you have issues. If you get serious about this and want
others to use it I'd recommend adding pester test cases which mock the
Appveyor/publish calls so that changes can be tested locally. I can help
you get started with that if you want.

 - Usage of ```{``` is not consistent
 - Instead of " use ' by default unless you have expressions you want to
    expand
 - Instead of ```'error' | Write-Host -ForegroundColor Red use 'error' |
   Write-Error```
 - use lowercase in [cmdletbinding()]
 - Vsix-PushArtifact can be improved the way it's authored it will hide errors. For ex if
 - publishToGallery is passed but there is no value for $filename
 - This script is specifically targeted at AppVeyor so Write-Host may be OK, but I'd recommend trying to avoid that in general. Instead use Write-Verbose/Write-Output/Write-Warning/Write-Error
 - If this is designed to be used with Import-Module then you should rename this file to be end in .psm1 so that it's obvious how to consume it